Split backend code into more files
